Add iscsi-initiator-utils to base profile
Source: MontaVista Software, LLC
MR: 63166
Type: Integration
Disposition: Local
Description:
iSCSI is part of base profile, and iscsi-initiator-utils  is the
endpoint that initiates a SCSI session.

Both open-iscsi-user and iscsi-initiator-utils both provide same
functionality. Preferring iscsi-initiator-utils as it is already
part of meta-networking layer, which is included in base profile

Reviewed-by: Jeremy Puhlman <jpuhlman@mvista.com>
Signed-off-by: Jagadeesh Krishnanjanappa <jkrishnanjanappa@mvista.com>