Make sure that permissions are correct in do_package
Source: MontaVista Software, LLC
MR: 66990
Type: Defect Fix
Disposition: Local
ChangeID: d877e3eca037142d6c4678bfd695813370ffc41a
Description:

Fix issues with interactions between pseudo and 32bit built
toolchain.

When the 32 bit build toolchain runs on a 64bit system the pseudo
libraries are not able to load unless pseudo has been built with
32bit support. On occasion pseudo loses track of the inode of the
file during strip and split, and rediscovers the inode is associated
with another file. It then thinks that this is a hard link and wont
let the value be set differently then the other file.

This patch tests to see if strip works with psuedo and if not moves
bits around so that the inodes are correct. It also fixes up a couple
other lingering owner/mode issues produced by the same interaction.

Signed-off-by: Jeremy Puhlman <jpuhlman@mvista.com>