image-groups: Do not deploy packagedata for kernel recipes
Source: MontaVista Software, LLC
MR: 71585
Type: Defect Fix
Disposition: Local
ChangeID: a839d794d1d5310c3aa9e12104b33c87d79492f6
Description:
Running SSTATE do_deploy_packagedata task on kernel recipes, will deploy
kernel pkgdata files at the same location, which are previously deployed
by other kernel recipes. This generates an error message as below:
-- snip --
ERROR: The recipe linux-mvista-mvtest-cgx-ph1 is trying to install files into a shared area when those files already exist. Those files and their manifest location are:
   .../tmp/deploy/pkgdata/runtime/kernel-src
 Matched in b'manifest-cavium-thunderx-linux-mvista.deploy_packagedata'
 .../tmp/deploy/pkgdata/runtime/kernel-src.packaged
 Matched in b'manifest-cavium-thunderx-linux-mvista.deploy_packagedata'
 .../tmp/deploy/pkgdata/runtime/kernel-map
 Matched in b'manifest-cavium-thunderx-linux-mvista.deploy_packagedata'
...
Please verify which recipe should provide the above files.
The build has stopped as continuing in this scenario WILL break things, if not now, possibly in the future (we've seen builds fail several months later). If the system knew how to recover from this automatically it would however there are several different scenarios which can result in this and we don't know which one this is. It may be you have switched providers of something like virtual/kernel (e.g. from linux-yocto to linux-yocto-dev), in that case you need to execute the clean task for both recipes and it will resolve this error. It may be you changed DISTRO_FEATURES from systemd to udev or vice versa. Cleaning those recipes should again resolve this error however switching DISTRO_FEATURES on an existing build directory is not supported, you should really clean out tmp and rebuild (reusing sstate should be safe). It could be the overlapping files detected are harmless in which case adding them to SSTATE_DUPWHITELIST may be the correct solution. It could also be your build is including two different conflicting versions of things (e.g. bluez 4 and bluez 5 and the correct solution for that would be to resolve the conflict. If in doubt, please ask on the mailing list, sharing the error and filelist above.
ERROR: If the above message is too much, the simpler version is you're advised to wipe out tmp and rebuild (reusing sstate is fine). That will likely fix things in most (but not all) cases.
-- snip --

Reviewed-by: Jeremy Puhlman <jpuhlman@mvista.com>
Signed-off-by: Jagadeesh Krishnanjanappa <jkrishnanjanappa@mvista.com>
Signed-off-by: Jeremy Puhlman <jpuhlman@mvista.com>