mvl.bbclass: use the KERNEL_CONF_LIST in do_kernel_postconfigure task
The KERNEL_CONF_LIST is pre-poluated when parsing and used in vardep of
the do_kernel_postconfigure so if they cahnge do_kernel_postconfigure
will run again.
It can be used in get_kernel_config_vars to retrieve the kernel configuration
present in bitbake environment without need to parse everithing again.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>